fix(flutter): Use NSNumber for Cocoa sample rate
Pass sampleRate through as NSNumber when configuring Sentry Cocoa options.
This matches the native API type and fixes the Swift compile error.

Fixes #3720

+++ packages/flutter/example/ios/RunnerTests/SentryFlutterTests.swift
@@ -58,7 +58,7 @@ final class SentryFlutterTests: XCTestCase {
     )
 
     XCTAssertEqual("https://e85b375ffb9f43cf8bdf9787768149e0@o447951.ingest.sentry.io/5428562", fixture.options.dsn)
-    XCTAssertEqual(0.25, fixture.options.sampleRate)
+    XCTAssertEqual(0.25, fixture.options.sampleRate.doubleValue)
     XCTAssertEqual(true, fixture.options.debug)
     XCTAssertEqual("fixture-environment", fixture.options.environment)
     XCTAssertEqual("fixture-release", fixture.options.releaseName)
