+++ packages/flutter/ios/sentry_flutter/Sources/sentry_flutter/SentryFlutter.swift
@@ -11,7 +11,7 @@ public final class SentryFlutter {
             options.dsn = dsn
         }
         if let sampleRate = data["sampleRate"] as? NSNumber {
-            options.sampleRate = sampleRate.floatValue
+            options.sampleRate = sampleRate
         }
         if let isDebug = data["debug"] as? Bool {
             options.debug = isDebug
